Rename backward.py to compatibility.py
